app/requirements.txt
Django~=3.2
django-admin-searchable-dropdown~=1.0
django-cors-headers~=3.7
django-dotenv~=1.4
djangorestframework~=3.12
djangorestframework-simplejwt~=5.2
gunicorn~=20.1
Jinja2~=3.1
Pillow~=9.0
psycopg2-binary
PyJWT~=2.0
pytz
PyYAML~=5.3
requests~=2.26
drf-yasg[validation]~=1.21
djangorestframework-camel-case
django-debug-toolbar~=3.2
django-imagekit
sentry-sdk~=1.29.2
django-filter~=23.02